# bulk_indexer: Use "create" constant